tests/wpt/web-platform-tests/css/cssom/ttwf-cssom-document-extension.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM - Extensions to the Document Interface: StyleSheetList length is 0 when no sheets loaded</title>
    <link rel="author" title="Jesse Bounds" href="mailto:jesse@codeforamerica.org">
    <link rel="reviewer" title="Ms2ger" href="mailto:ms2ger@gmail.com"> <!-- 2012-06-17 -->
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#extensions-to-the-document-interface">
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-stylesheetlist-interface">
    <meta name="flags" content="dom">
    <meta name="assert" content="The styleSheets attribute must return a StyleSheetList sequence representing the document style sheets.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="log"></div>
    <script>
        // Get the Document's styleSheets attribute
        var styleSheets = document.styleSheets;
        // Verify that the styleSheets list length is 0 because no stylesheets have been loaded
        test(function() {
          assert_equals(styleSheets.length, 0, "styleSheets.length is incorrect:");
        })
    </script>
</body>
</html>
